refactor(webapp): split agentPanel into stream, controls, and skill formatting
agentPanel.js had grown to 1291 lines while every sibling in js/agent/ sits
between 5 and 12KB, doing roster, chat, skill grouping and composer at once.

Three modules come out of it:

  skillFormat.js      pure naming/arg formatting + skill-card DOM builders
  chatStream.js       the transcript and all its state
  directiveControls.js  agent picker, Start/Stop, Reset

The transcript stays whole rather than splitting further: scroll pinning, the
open thought group, live skill runs and the streak are mutually entangled — one
message can touch all four — so threading them between modules would add more
coupling than it removes. agentPanel keeps the shell, composer and ROS wiring,
and is now 324 lines.

Three things are not pure movement:

  - The active-skill chip belongs to the control panel, not the transcript, so
    chatStream reports the running skill out through onActiveSkill and tracks
    the name itself instead of reading it back off textContent.
  - The live chat_out handler and replayEntry had identical routing chains;
    they share routeChatOut now so the two paths cannot drift.
  - History is replay(entries) rather than loadHistory(): the service call
    stays with the rest of the ROS wiring, which also makes the transcript
    testable from a fixture.

index.html preloads the three new modules — tests/preload.test.js catches the
extra round trips otherwise.
