+++ webapp/index.html
@@ -32,6 +32,9 @@
     <link rel="modulepreload" href="/js/micAudioState.js" />
     <link rel="modulepreload" href="/js/agent/main.js" />
     <link rel="modulepreload" href="/js/agent/agentPanel.js" />
+    <link rel="modulepreload" href="/js/agent/chatStream.js" />
+    <link rel="modulepreload" href="/js/agent/directiveControls.js" />
+    <link rel="modulepreload" href="/js/agent/skillFormat.js" />
     <link rel="modulepreload" href="/js/agent/challengePanel.js" />
     <link rel="modulepreload" href="/js/agent/challengeIntro.js" />
     <link rel="modulepreload" href="/js/agent/micStream.js" />
